moved colorimetrie directory
