+++ src/colorimetrie/colorimetrie.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 #include <err.h>
 #include "colorimetrie.h"
-#include "../image.h"
+#include "../image/image.h"
 #include "stack.h"
 #include <math.h>
