+++ src/colorimetrie/images/results/rectangle
