+++ src/colorimetrie/stack.h
